refactor how to modals
